backend/requirements.txt
fastapi[standard]
pydantic==2.5.0
pydantic-settings==2.1.0
azure-cosmos==4.5.0
python-multipart
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart